refactor: Refactor a piece of cake
