+++ todomon-core/src/main/java/com/maruhxn/todomon/core/domain/item/implement/ItemRemover.java
@@ -11,7 +11,5 @@ public class ItemRemover {
 
     private final ItemRepository itemRepository;
 
-    public void remove(Item item) {
-        itemRepository.delete(item);
-    }
+
 }
