+++ todomon-core/src/main/java/com/maruhxn/todomon/core/domain/item/implement/ItemWriter.java
@@ -7,11 +7,15 @@ import org.springframework.stereotype.Component;
 
 @Component
 @RequiredArgsConstructor
-public class ItemCreator {
+public class ItemWriter {
 
     private final ItemRepository itemRepository;
 
     public void create(Item item) {
         itemRepository.save(item);
+    }
+
+    public void remove(Item item) {
+        itemRepository.delete(item);
     }
 }
