+++ todomon-core/src/main/java/com/maruhxn/todomon/core/domain/member/application/TitleNameService.java
@@ -1,11 +1,11 @@
 package com.maruhxn.todomon.core.domain.member.application;
 
 import com.maruhxn.todomon.core.domain.member.domain.Member;
+import com.maruhxn.todomon.core.domain.member.domain.TitleName;
 import com.maruhxn.todomon.core.domain.member.dto.request.UpsertTitleNameReq;
 import com.maruhxn.todomon.core.domain.member.implement.MemberReader;
-import com.maruhxn.todomon.core.domain.member.implement.TitleNameCreator;
 import com.maruhxn.todomon.core.domain.member.implement.TitleNameReader;
-import com.maruhxn.todomon.core.domain.member.implement.TitleNameRemover;
+import com.maruhxn.todomon.core.domain.member.implement.TitleNameWriter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -16,20 +16,21 @@ import org.springframework.transaction.annotation.Transactional;
 public class TitleNameService {
 
     private final MemberReader memberReader;
-    private final TitleNameCreator titleNameCreator;
     private final TitleNameReader titleNameReader;
-    private final TitleNameRemover titleNameRemover;
+    private final TitleNameWriter titleNameWriter;
 
     public void upsertTitleName(Member member, UpsertTitleNameReq req) {
         titleNameReader.findByMember_Id(member.getId())
                 .ifPresentOrElse(
                         tn -> tn.update(req.getName(), req.getColor()),
-                        () -> titleNameCreator.create(member, req)
+                        () -> titleNameWriter.create(member, req)
                 );
     }
 
     public void deleteTitleName(Long memberId) {
         Member member = memberReader.findById(memberId);
-        titleNameRemover.delete(member);
+        TitleName titleName = titleNameReader.findByMemberId(member.getId());
+        member.setTitleName(null);
+        titleNameWriter.delete(titleName);
     }
 }
